test(itest): move tx voiding tests to fullnode-specific
Migrate the 'processing transaction metadata changes' describe block
out of hathorwallet_others.test.ts into a new fullnode-specific file.

Voiding is simulated by capturing the websocket IHistoryTx payload via
a spy on HathorWallet.onNewTx, flipping is_voided, and re-injecting it,
then asserting on MemoryStore internals. That machinery is not exposed
by the wallet-service facade, so the cases stay fullnode-specific.

Remove the now-unused lodash, MemoryStore, IHistoryTx and HathorWallet
imports left behind in the source file.
